AVRO-3912: [Rust] Extract the code for BigDecimal to its own module (bigdecimal.rs)
